separated collectionsTasks test from collections tests for organization
